@@ -89,53 +95,53 @@
     "babel-preset-react": "^6.24.1",
     "babel-preset-stage-2": "^6.24.1",
     "chokidar": "^1.6.1",
+    "concurrently": "^4.0.1",
     "cpx": "^1.3.2",
     "cross-env": "^4.0.0",
     "electron-builder": "^11.2.4",
     "electron-builder-squirrel-windows": "^11.2.1",
-    "electron-devtools-installer": "^2.2.0",
+    "electron-devtools-installer": "^2.2.4",
     "emojione": "^2.2.7",
-    "eslint": "^3.14.0",
+    "eslint": "^5.6.0",
     "eslint-config-google": "^0.7.1",
-    "eslint-plugin-babel": "^4.1.1",
-    "eslint-plugin-flowtype": "^2.30.0",
-    "eslint-plugin-react": "^7.4.0",
+    "eslint-plugin-babel": "^4.1.2",
+    "eslint-plugin-flowtype": "^2.50.3",
+    "eslint-plugin-react": "^7.11.1",
     "expect": "^1.16.0",
     "fs-extra": "^0.30.0",
     "html-webpack-plugin": "^3.2.0",
     "json-loader": "^0.5.3",
-    "karma": "^1.7.0",
+    "karma": "^3.0.0",
     "karma-chrome-launcher": "^0.2.3",
-    "karma-cli": "^0.1.2",
-    "karma-junit-reporter": "^0.4.1",
+    "karma-cli": "^1.0.1",
+    "karma-junit-reporter": "^2.0.0",
     "karma-logcapture-reporter": "0.0.1",
-    "karma-mocha": "^0.2.2",
+    "karma-mocha": "^1.3.0",
     "karma-sourcemap-loader": "^0.3.7",
     "karma-spec-reporter": "0.0.31",
-    "karma-summary-reporter": "^1.3.3",
-    "karma-webpack": "^v4.0.0-beta.0",
+    "karma-summary-reporter": "^1.5.1",
+    "karma-webpack": "4.0.0-beta.0",
     "matrix-mock-request": "^1.2.0",
     "matrix-react-test-utils": "^0.2.0",
     "minimist": "^1.2.0",
     "mkdirp": "^0.5.1",
-    "mocha": "^2.4.5",
-    "parallelshell": "^3.0.2",
+    "mocha": "^5.2.0",
     "postcss-extend": "^1.0.5",
     "postcss-import": "^11.1.0",
-    "postcss-loader": "^2.1.4",
+    "postcss-loader": "^2.1.6",
     "postcss-mixins": "^6.2.0",
     "postcss-nested": "^3.0.0",
-    "postcss-scss": "^1.0.5",
+    "postcss-scss": "^1.0.6",
     "postcss-simple-vars": "^4.1.0",
     "postcss-strip-inline-comments": "^0.1.5",
     "raw-loader": "^0.5.1",
     "react-addons-perf": "^15.4.0",
     "react-addons-test-utils": "^15.6.0",
     "rimraf": "^2.4.3",
-    "source-map-loader": "^0.2.3",
-    "webpack": "^4.12.0",
-    "webpack-cli": "^3.0.3",
-    "webpack-dev-server": "^3.1.4"
+    "source-map-loader": "^0.2.4",
+    "webpack": "^4.20.2",
+    "webpack-cli": "^3.1.1",
+    "webpack-dev-server": "^3.1.9"
   },
   "optionalDependencies": {
     "olm": "https://matrix.org/packages/npm/olm/olm-2.2.1.tgz"
